model NfEntrada: qtd passa a ser inteiro

+++ src/logistica/models.py
@@ -169,7 +169,7 @@ class NfEntrada(models.Model):
     emissor = models.CharField(max_length=200)
     numero = models.IntegerField("número")
     descricao = models.CharField("descrição", max_length=200)
-    qtd = models.DecimalField("quantidade", max_digits=13, decimal_places=4)
+    qtd = models.IntegerField("quantidade")
     chegada = models.DateTimeField("chegada", default=timezone.now)
     transportadora = models.CharField(max_length=100)
     motorista = models.CharField(max_length=100)
